Move environment variables related functions/definitions into env.{c,h} without any further changes
